renamed slices to ducks
