+++ src/features/lectures/AddLectureButton.tsx
@@ -2,7 +2,7 @@ import { Button } from '@chakra-ui/button';
 import { AddIcon } from '@chakra-ui/icons';
 import React, {Dispatch, SetStateAction} from 'react';
 import { useAppDispatch } from '../../app/hooks';
-import { addLecture } from './lectureSlice';
+import { addLecture } from './lectureDuck';
 
 const AddLectureButton = () => {
   const dispatch = useAppDispatch();
